requirements.txt
asyncpg=0.27.0 
discord.py=2.2.3 